AG-14830 - Rename folder
