+++ packages/ag-charts-enterprise/src/features/band-highlight/bandHighlightModule.ts
