+++ packages/ag-charts-enterprise/src/features/band-highlight/main.ts
